+++ actions/charts/charts.go
@@ -12,7 +12,6 @@ const (
 	rancherChartsName    = "rancher-charts"
 	rancherPartnerCharts = "rancher-partner-charts"
 	active               = "active"
-	metadataName         = "metadata.name="
 )
 
 // InstallOptions is a struct of the required options to install a chart.
